+++ NEWS.md
@@ -4,6 +4,9 @@
 
 * Disambiguates the `verbose` and `verbose_iter` control options to better align with documented functionality. The former controls logging for general progress updates, while the latter only does so for the Bayesian search process. (#682) 
 
+* `last_fit()` will now error when supplied a fitted workflow. (#678)
+
+
 # tune 1.1.1
 
 * Fixed a bug introduced in tune 1.1.0 in `collect_()` functions where the